Now with rank change class
the rank change class tracks the as it changes over time.
the issue I was having was that updating the rank was unreliable.
it would work sometimes and not others.

the idea with this approach is to build the leader board from the
querying the rank changes and getting teh current rank for each member as it has changed
over time.

tests :passing